Move request rate limiting into mcp
Constraint: SessionRateLimiter is only used to protect MCP/HTTP request handling and tenant request buckets.\nRejected: Keeping rate-limiter under src/utils | it is protocol runtime policy, not a generic leaf helper.\nRejected: Adding a compatibility shim | all in-repo callers were updated directly and no package export exposes this path.\nConfidence: high\nScope-risk: narrow\nDirective: Keep MCP request accounting and throttling under src/mcp.\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/mcp/rate-limiter-sweep.test.ts tests/integration/multi-tenant-auth.test.ts tests/mcp/server.test.ts tests/mcp/server-tenant-metrics.test.ts\nTested: npm run build\nTested: npm run lint\nTested: npm run lint:tier\nTested: npm run lint:repo-structure\nTested: git diff --check\nNot-tested: Full Jest suite.
